@@ -2,9 +2,9 @@ package com.axzae.homeassistant.fragment;
 
 import android.app.Activity;
 import android.app.Application;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.axzae.homeassistant.AppController;
 import com.axzae.homeassistant.MainActivity;
