@@ -16,7 +16,7 @@
 
 package com.axzae.homeassistant.helper;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Listener for manual initiation of a drag.
