@@ -12,7 +12,7 @@ import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Binder;
 import android.os.IBinder;
-import android.support.v4.app.NotificationCompat;
+import androidx.core.app.NotificationCompat;
 import android.util.Log;
 import android.widget.Toast;
 
@@ -30,7 +30,6 @@ import com.axzae.homeassistant.util.CommonUtil;
 import org.json.JSONObject;
 
 import java.util.HashMap;
-import java.util.Random;
 
 import io.reactivex.subjects.PublishSubject;
 import io.reactivex.subjects.Subject;
