@@ -1,6 +1,6 @@
 package com.axzae.homeassistant.shared;
 
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 
 public interface OnBackPressedInterface {
     /**
